+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - `_tee` now supports both "err" and "out" [#215](https://github.com/amoffat/sh/issues/215)
 - Removed the builtin override `cd` [link](https://github.com/amoffat/sh/pull/584#discussion_r698055681)
 - Altered process launching model to behave more expectedly [#495](https://github.com/amoffat/sh/issues/495)
+- Bugfix where `_no_out` isn't allowed with `_iter="err"` [#638](https://github.com/amoffat/sh/issues/638)
 
 ## 1.14.3 - 7/17/22